adding changes and swagger

+++ src/main/java/com/example/demo/Impl/CourseServiceImpl.java
@@ -1,9 +1,12 @@
 package com.example.demo.Impl;
 
 import com.example.demo.Services.CourseService;
+import com.example.demo.dto.CourseDTO;
 import com.example.demo.error.DemoProjectException;
 import com.example.demo.model.Course;
 import com.example.demo.repository.CourseRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,22 +19,33 @@ import static com.example.demo.error.ErrorInterface.ErrorEnum;
 @Service
 public class CourseServiceImpl implements CourseService {
 
+    public static final Logger LOGGER= LoggerFactory.getLogger(CourseService.class);
     @Autowired
     private CourseRepository courseRepository;
 
     @Override
-    public Course saveCourse(Course course) {
-        return courseRepository.save(course);
+    public CourseDTO saveCourse(CourseDTO courseDTO) {
+        LOGGER.info("Entering saveCourse service");
+        Course course=new Course();
+        course.setCourseName(courseDTO.getCourseName());
+        course.setCredits(courseDTO.getCredits());
+        return new CourseDTO(courseRepository.save(course));
     }
 
     @Override
     public List<Course> fetchCourses() {
+        LOGGER.info("Entering fetchCourses service");
         return (List<Course>) courseRepository.findAll();
     }
 
     @Override
     public Course updateCourse(Course course, Long courseId) {
-        Course currentCourse = courseRepository.findById(courseId).get();
+        Optional<Course> optionalCourse = courseRepository.findById(courseId);
+        if (!optionalCourse.isPresent()) {
+            throw new DemoProjectException(ErrorEnum.ID_NOT_FOUND.getHttpStatus(), ErrorEnum.ID_NOT_FOUND.getErrorMessage());
+        }
+
+        Course currentCourse = optionalCourse.get();
 
         if (Objects.nonNull(course.getCourseName()) && !"".equalsIgnoreCase(course.getCourseName())) {
             currentCourse.setCourseName(course.getCourseName());
@@ -42,6 +56,10 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public void deleteCourseById(Long courseId) {
+        Optional<Course> course = courseRepository.findById(courseId);
+        if (!course.isPresent()) {
+            throw new DemoProjectException(ErrorEnum.ID_NOT_FOUND.getHttpStatus(), ErrorEnum.ID_NOT_FOUND.getErrorMessage());
+        }
         courseRepository.deleteById(courseId);
     }
